Write the specifications for file finders
`FileFinder` and `FilesFinder` interfaces were defined along with
their asynchronous and synchronous definitions. Tests were written
in order to ensure the proper implementation of their features as
described in their documentation.
A `Matcher` function type was defined as a way to test files that
are to be considered among the found files after the execution of
the aforementioned functions. A basic `Matcher` function,
`ofBasename`, was implemented and tested for the specification
tests of file finders.