Implement `findFile` and `findFiles` functions
The functions `findFile` and `findFiles` were implemented in accordance with the previously implemented interfaces. However, support for `URL` and `Buffer` paths has been dropped, because resolving paths using a `URL` with the `file:` protocol involves additional comparisons which are not needed for standard string paths.
Tests were added for both the `findFile` and `findFiles` functions to ensure maximum coverage.
Additional optimizations were made.