Implement `findFirstFile` function
A `FirstFileFinder` interface was added based on the `FindFile` interface. This new interface has the same functionality as `FindFile`, however it does not throw if there is more than one file that matches the sequence of tests; rather it returns the first one it encounters that passes the tests.
This new function provides an optimization over the standard `findFile` function for cases where the sequence of tests has a uniqueness factor, meaning that there can only exist one path in a directory that passes all the tests. For instance, searches for a file of a given base name are best executed with the new function.