Fix asynchronous file finders
The file finders were renamed to better reflect their functionality.
`findFiles` is now `findAllFiles`, `findFile` is now `strictFindFile` and `findFirstFile` is now `findFile`.
Test cases were added for edge cases, and they were reorganized per function signature.
The previous test coverage revealed that the body of the asynchronous implementation of the file finders were not being executed. This is because Mocha promise assertions requires the promise to be returned. Properly running the tests revealed that none of the implementations were functioning as intended. This has been corrected by using the synchronous `readdir` function.