Add path extractor matcher functions
Added matcher functions for the directory and extension name of a path using `path.dirname` and `path.extname`.
The body of `ofBasename` has been generalized and moved over to the `match.ts` file. This generalization supports different path part extractors and a strategy function for chaining the various tests that are performed on the extracted part.
With these changes, it is now possible to search for a file whose base name matches a regular expression omitting the extension name, and matches with a set of extension names.