Use iterator pattern and add matcher functions
Functions were added to allow for the usage of iterables in file-finding functions. These allow for iterables to be concatenated, filtered and mapped as if they were arrays, which does slow down certain features for small directories, but greatly improves their overall performance for large directories.
Also, these iterable operations will allow for the implementation of iterative downward traversals of the file system. 
The definition of a matcher function has been revised to be generic.
Matcher functions were added in order to facilitate certain precise queries on the file system.