Add downward directories fetcher
A rudimentary downward directories fetcher has been implemented to allow for the searching of files downwards in the file system graph.
The iterator pattern has been added to asynchronous file finders.
The TypeScript configuration has been altered in order to use the most recent JavaScript library typing, hence allowing for the removal of unnecessary type casting in the return object of iterator results.
However, moving forward, the code base will get rid of the currently implemented iterator pattern and instead use generators. These will reduce the cognitive complexity of functions as well as making the code more readable. Moreover, the tests will be reviewed in order to be truly unit-based.