Revise the implementation of file finders
File finders and their tests have been revised with the implementation of iterable readdirs. Hybrid types are no longer used to have the synchronous file finders as an attribute of the asynchronous file finder. This reduces the repetition of code as well as make it more readable. Moreover, the new format respects the convention set by Node in their libraries.