Revise the implementation of `hasFile` filter
The `hasFile` filter has been revised to use the new iterable solutions for `readdir` and to allow for an asynchronous `hasfile` function.
The `mappedIterable` and `filteredIterable` functions have been successfully removed. `matchingPathsIterator` and `matcher` files remain.
A path segment iterator has been implemented to allow for the reintroduction of a synchronous filter over the segments of a path.
The development dependency `chai-as-promised` has been dropped and replaced with the usage of the `rejects` function from the standard assertion library.